feat: 멤버 관련 커스텀 에러 클래스 패키지 위치 변경
