--- src/main/java/econovation/moongtaengi/global/exception/AlreadyRegisteredException.java
@@ -1,8 +1,0 @@
-package econovation.moongtaengi.global.exception;
-
-public class AlreadyRegisteredException extends BusinessException {
-
-    public AlreadyRegisteredException() {
-        super(ErrorCode.ALREADY_REGISTERED);
-    }
-}
