--- src/main/java/econovation/moongtaengi/global/exception/MemberNotFoundException.java
@@ -1,8 +1,0 @@
-package econovation.moongtaengi.global.exception;
-
-public class MemberNotFoundException extends BusinessException {
-
-    public MemberNotFoundException() {
-        super(ErrorCode.MEMBER_NOT_FOUND);
-    }
-}
