+++ src/main/java/econovation/moongtaengi/member/exception/MemberNotFoundException.java
@@ -1,0 +1,11 @@
+package econovation.moongtaengi.member.exception;
+
+import econovation.moongtaengi.global.exception.BusinessException;
+import econovation.moongtaengi.global.exception.ErrorCode;
+
+public class MemberNotFoundException extends BusinessException {
+
+    public MemberNotFoundException() {
+        super(ErrorCode.MEMBER_NOT_FOUND);
+    }
+}
